refactor: move into use cases
