+++ src/features/entity/use_cases/instantiate_entity_use_case/instantiate_entity_use_case.py
@@ -1,5 +1,7 @@
 from authentication.models import User
-from common.entity.create_entity import CreateEntity
+from features.entity.use_cases.instantiate_entity_use_case.create_entity import (
+    CreateEntity,
+)
 from restful.request_types.shared import Entity
 from services.document_service.document_service import DocumentService
