+++ src/features/entity/use_cases/validate_entity_use_case.py
