+++ src/features/entity/use_cases/validate_existing_entity_use_case.py
