+++ src/tests/unit/common/utils/test_create_entity.py
@@ -1,7 +1,9 @@
 import unittest
 
-from common.entity.create_entity import CreateEntity
 from domain_classes.blueprint_attribute import BlueprintAttribute
+from features.entity.use_cases.instantiate_entity_use_case.create_entity import (
+    CreateEntity,
+)
 from tests.unit.mock_data.mock_blueprint_provider import MockBlueprintProvider
 from tests.unit.mock_data.mock_document_service import get_mock_document_service
